refactor: create layouts directory, split original layouts file, extract methods
